Add data to endpoint
Add data to endpoint, fix imports, started to discount taxes from previous months
