+++ samples2/.~lock.operacoes.xls#
@@ -1,1 +1,1 @@
-,hda,mdm,03.02.2020 21:45,file:///home/hda/.config/libreoffice/4;
+,nathan,nathan-Notebook,02.03.2020 20:29,file:///home/nathan/snap/libreoffice/163/.config/libreoffice/4;
